VP-1677: Separate AccountOrderDetails.vue to vue and ts files
