+++ client-app/src/pages/account/components/order-details/order-details-item/index.vue
@@ -16,15 +16,4 @@
   </div>
 </template>
 
-<script lang="ts">
-import Vue from "vue";
-import Component from "vue-class-component";
-import { Prop } from "vue-property-decorator";
-import { OrderLineItem } from "@common/api/api-clients";
-
-@Component
-export default class OrderDetailsItem extends Vue {
-  @Prop()
-  item!: OrderLineItem;
-}
-</script>
+<script lang="ts" src="./order-details-item"></script>
