+++ client-app/src/pages/account/components/order-details/order-details-item/order-details-item.ts
@@ -1,0 +1,10 @@
+import Vue from "vue";
+import Component from "vue-class-component";
+import { Prop } from "vue-property-decorator";
+import { OrderLineItem } from "@common/api/api-clients";
+
+@Component
+export default class OrderDetailsItem extends Vue {
+  @Prop()
+  item!: OrderLineItem;
+}
