+++ client-app/src/pages/account/components/order-details/order-details-totals/order-details-totals.ts
@@ -1,0 +1,11 @@
+import Vue from "vue";
+import Component from "vue-class-component";
+import { Prop } from "vue-property-decorator";
+import { CustomerOrder } from "@common/api/api-clients";
+
+@Component
+export default class OrderDetailsTotals extends Vue {
+  @Prop()
+  order!: CustomerOrder;
+
+}
